n5 management

+++ src/main/java/com/bigdistributer/dataexchange/s3/func/BucketManager.java
@@ -6,6 +6,7 @@ import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.bigdistributer.dataexchange.utils.log.Log;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -35,15 +36,22 @@ public class BucketManager {
         return false;
     }
 
-    public void upload(File file) {
-        s3client.putObject(name, file.getName(), file);
+    public void upload(File file) throws IOException {
+        if (!file.exists())
+            throw new IOException("File not exist !");
+
+        if (file.isDirectory())
+            throw new RuntimeException("Not implimented yet !");
+
+        if (file.isFile())
+            s3client.putObject(name, file.getName(), file);
         Log.info("File " + file.getName() + " sent!");
     }
 
     public void upload(List<File> files) {
         List<PutObjectRequest> requestList = new ArrayList<>();
         for (File f : files)
-            requestList.add(new PutObjectRequest(name, "f/"+f.getName(), f));
+            requestList.add(new PutObjectRequest(name, "f/" + f.getName(), f));
         for (PutObjectRequest request : requestList) {
             s3client.putObject(request);
             Log.info("File " + request.getFile().getName() + " sent!");
